refactor: Extract AgentDetail as reusable component
## New Component: AgentDetail
- Extracted from /agents/[agentId]/page.tsx into shared component
- Props: agent, onUpdate, onBack, backLabel, compact
- Also exports: AgentDetailData type, AgentDetailSkeleton, AgentDetailNotFound

## Changes to /agents/[agentId]/page.tsx
- Now uses AgentDetail component instead of inline code
- Reduced from 424 lines to ~95 lines

## Changes to Command Center Agents tab
- Clicking agent card now shows AgentDetail inline (no navigation)
- Back button returns to agent grid view
- Stays within Command Center context
